Bench using criterion
